Create leastNumberDict
This Python program takes a file with lines of integers 0-9
and returns the number that occurs the least amount of times in
the file. The program also returns a summary in the form of a
dictionary with keys being an integer 0-9 and the values
are the total count of that integer. Replace "numbers.txt" filename with file path name of your file.